Use ChannelBuilder created by ThingHandlerCallback
