+++ bundles/org.openhab.binding.deconz/src/main/java/org/openhab/binding/deconz/internal/handler/LightThingHandler.java
@@ -46,8 +46,7 @@ import org.openhab.core.thing.Thing;
 import org.openhab.core.thing.ThingStatus;
 import org.openhab.core.thing.ThingStatusDetail;
 import org.openhab.core.thing.ThingTypeUID;
-import org.openhab.core.thing.binding.builder.ChannelBuilder;
-import org.openhab.core.thing.binding.builder.ThingBuilder;
+import org.openhab.core.thing.type.ChannelKind;
 import org.openhab.core.types.Command;
 import org.openhab.core.types.CommandOption;
 import org.openhab.core.types.RefreshType;
@@ -346,22 +345,13 @@ public class LightThingHandler extends DeconzBaseThingHandler {
         }
 
         ChannelUID effectChannelUID = new ChannelUID(thing.getUID(), CHANNEL_EFFECT);
-        ChannelUID effectSpeedChannelUID = new ChannelUID(thing.getUID(), CHANNEL_EFFECT_SPEED);
-
-        if (thing.getChannel(CHANNEL_EFFECT) == null) {
-            ThingBuilder thingBuilder = editThing();
-            thingBuilder.withChannel(
-                    ChannelBuilder.create(effectChannelUID, "String").withType(CHANNEL_EFFECT_TYPE_UID).build());
-            if (model == EffectLightModel.LIDL_MELINARA) {
-                // additional channels
-                thingBuilder.withChannel(ChannelBuilder.create(effectSpeedChannelUID, "Number")
-                        .withType(CHANNEL_EFFECT_SPEED_TYPE_UID).build());
-            }
-            updateThing(thingBuilder.build());
-        }
+        createChannel(CHANNEL_EFFECT, ChannelKind.STATE);
 
         switch (model) {
             case LIDL_MELINARA:
+                // additional channels
+                createChannel(CHANNEL_EFFECT_SPEED, ChannelKind.STATE);
+
                 List<String> options = List.of("none", "steady", "snow", "rainbow", "snake", "tinkle", "fireworks",
                         "flag", "waves", "updown", "vintage", "fading", "collide", "strobe", "sparkles", "carnival",
                         "glow");
