+++ bundles/org.openhab.binding.deconz/src/main/java/org/openhab/binding/deconz/internal/handler/SensorThermostatThingHandler.java
@@ -39,9 +39,7 @@ import org.openhab.core.library.types.StringType;
 import org.openhab.core.thing.ChannelUID;
 import org.openhab.core.thing.Thing;
 import org.openhab.core.thing.ThingTypeUID;
-import org.openhab.core.thing.binding.builder.ChannelBuilder;
-import org.openhab.core.thing.binding.builder.ThingBuilder;
-import org.openhab.core.thing.type.ChannelTypeUID;
+import org.openhab.core.thing.type.ChannelKind;
 import org.openhab.core.types.Command;
 import org.openhab.core.types.RefreshType;
 import org.slf4j.Logger;
@@ -201,11 +199,8 @@ public class SensorThermostatThingHandler extends SensorBaseThingHandler {
 
         SensorMessage sensorMessage = (SensorMessage) stateResponse;
         SensorState sensorState = sensorMessage.state;
-        if (sensorState != null && sensorState.windowopen != null && thing.getChannel(CHANNEL_WINDOWOPEN) == null) {
-            ThingBuilder thingBuilder = editThing();
-            thingBuilder.withChannel(ChannelBuilder.create(new ChannelUID(thing.getUID(), CHANNEL_WINDOWOPEN), "String")
-                    .withType(new ChannelTypeUID(BINDING_ID, "open")).build());
-            updateThing(thingBuilder.build());
+        if (sensorState != null && sensorState.windowopen != null) {
+            createChannel(CHANNEL_WINDOWOPEN, ChannelKind.STATE);
         }
 
         super.processStateResponse(stateResponse);
